MNT: Rename some q2-qurro internals to be clearer
Now that Qarcoal is around, having filenames like "_method.py" lying
around will be pretty confusing. This should make it easier to
maintain the codebase in the future.
